Session Timer
Added a timer which tracks the whole session length, exports when application is quit. Formatted in hrs,mins,secs,ms.

Might want to check the order in which application quit method is called as it is possible the one for analytics export may be executed before the session timer sends it to the anlytics.